Moving modules to remove _new and updating references
